requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
openai==0.27.0
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
joblib==1.1.0
python-dotenv==0.19.2

python deploy.py
env\Scripts\activate
http://localhost:5000